requirements.txt
openai
pandas
pydantic
python-dotenv
